services/agd-shared/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "agd-shared"
version = "0.1.0"
description = "Shared JWT auth, classification enforcement, and error helpers for AGD Python services"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111.0",
    "python-jose[cryptography]>=3.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2",
    "httpx>=0.27",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["agd_shared*"]
